Temporal Guard/pyproject.toml
[project]
name = "temporalguard"
version = "0.1.0"
description = "Runtime Verification for LLM Conversations using Past-Time Temporal Logic"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "aiosqlite>=0.20",
    "httpx>=0.27",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "playwright>=1.42",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]

[tool.coverage.run]
branch = true
source = ["backend"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "S",      # flake8-bandit
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line length handled by formatter; allow long strings/URLs
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",     # assert allowed in tests
    "PLR2004",  # magic values allowed in tests
    "ANN",      # annotations optional in tests
    "B017",     # pytest.raises(Exception) acceptable in tests
    "N806",     # uppercase variables (F, T, E) used as formula/trace/expected
    "RUF002",   # Unicode en-dash in docstrings (ptLTL notation)
    "RUF003",   # Unicode en-dash in comments (ptLTL notation)
]
"scripts/**/*.py" = [
    "T201",     # print allowed in CLI scripts
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
